Reconfigure test into libraryies and nodes with unit tests
