+++ tesseract_ros/tesseract_examples/tesseract_ros_examples/launch/pick_and_place_example.launch
@@ -7,6 +7,8 @@
     <arg name="box_x" default="0.15"/>
     <arg name="box_y" default="0.4"/>
     <arg name="box_parent_link" default="workcell_base"/>
+    <arg name="rviz" default="true"/>
+    <arg name="testing" default="false"/>
 
     <!-- Load Robot Model    -->
     <param name="robot_description"  command="$(find xacro)/xacro --inorder $(find tesseract_ros_examples)/urdf/pick_and_place_plan.xacro"/>
@@ -25,16 +27,18 @@
     <param name="box_x" value="$(arg box_x)"/>
     <param name="box_y" value="$(arg box_y)"/>
     <param name="box_parent_link" value="$(arg box_parent_link)"/>
+    <group unless="$(arg testing)">
+      <!-- Launch planner node -->
+      <node name="pick_and_place_plan_node" pkg="tesseract_ros_examples" type="tesseract_ros_examples_pick_and_place_plan" output="screen">
+          <param name="steps_per_phase" value="$(arg steps_per_phase)"/>
+          <param name="plotting" value="$(arg plotting)"/>
+          <param name="file_write_cb" value="$(arg file_write_cb)" />
+          <param name="rviz" type="bool" value="$(arg rviz)"/>
+      </node>
 
-    <!-- Launch planner node -->
-    <node name="pick_and_place_plan_node" pkg="tesseract_ros_examples" type="tesseract_ros_examples_pick_and_place_plan" output="screen" required="true">
-        <param name="steps_per_phase" value="$(arg steps_per_phase)"/>
-        <param name="plotting" value="$(arg plotting)"/>
-        <param name="file_write_cb" value="$(arg file_write_cb)" />
-    </node>
-
-    <!-- Launch visualization -->
-    <node if="$(arg plotting)" pkg="rviz" type="rviz" name="pick_and_place_viz"
-        args="-d $(find tesseract_ros_examples)/config/examples.rviz" />
+      <!-- Launch visualization -->
+      <node if="$(arg rviz)" pkg="rviz" type="rviz" name="pick_and_place_viz"
+          args="-d $(find tesseract_ros_examples)/config/examples.rviz" />
+    </group>
 
 </launch>
