+++ tesseract_ros/tesseract_examples/tesseract_ros_examples/launch/puzzle_piece_example.launch
@@ -4,6 +4,8 @@
   <!-- The name of the parameter under which the URDF is loaded -->
   <arg name="robot_description" default="robot_description"/>
   <arg name="plotting" default="true"/>
+  <arg name="rviz" default="true"/>
+  <arg name="testing" default="false"/>
 
   <!-- Load universal robot description format (URDF) -->
   <param name="$(arg robot_description)" command="$(find xacro)/xacro --inorder '$(find tesseract_ros_examples)/urdf/puzzle_piece_workcell.xacro'" />
@@ -21,15 +23,18 @@
     <rosparam command="load" file="$(find tesseract_ros_examples)/config/kinematics.yaml"/>
   </group>
 
-  <!-- If needed, broadcast static tf for robot root -->
-  <node pkg="tf" type="static_transform_publisher" name="virtual_joint_broadcaster_0" args="0 0 0 0 0 0 world base_link 100" />
+  <group unless="$(arg testing)">
+    <!-- If needed, broadcast static tf for robot root -->
+    <node pkg="tf" type="static_transform_publisher" name="virtual_joint_broadcaster_0" args="0 0 0 0 0 0 world base_link 100" />
 
-  <node pkg="tesseract_ros_examples" type="tesseract_ros_examples_puzzle_piece_auxillary_axes_plan" name="tesseract_ros_examples_puzzle_piece_auxillary_axes_plan" required="true">
-    <param name="plotting" type="bool" value="$(arg plotting)"/>
-  </node>
+    <node pkg="tesseract_ros_examples" type="tesseract_ros_examples_puzzle_piece_example_node" name="tesseract_ros_examples_puzzle_piece_example_node">
+      <param name="plotting" type="bool" value="$(arg plotting)"/>
+      <param name="rviz" type="bool" value="$(arg rviz)"/>
+    </node>
 
-  <!-- Launch visualization -->
-  <node if="$(arg plotting)" pkg="rviz" type="rviz" name="tesseract_ros_examples_puzzle_piece_auxillary_axes_plan_rviz"
-      args="-d $(find tesseract_ros_examples)/config/examples.rviz" />
+    <!-- Launch visualization -->
+    <node if="$(arg rviz)" pkg="rviz" type="rviz" name="tesseract_ros_examples_puzzle_piece_example_rviz"
+        args="-d $(find tesseract_ros_examples)/config/examples.rviz" />
+  </group>
 
 </launch>
