+++ tesseract_ros/tesseract_examples/tesseract_ros_examples/package.xml
@@ -24,6 +24,8 @@
   <depend>octomap_ros</depend>
   <depend>pcl_conversions</depend>
 
+  <test>rostest</test>
+
   <export>
   </export>
 </package>
