+++ tesseract_ros/tesseract_examples/tesseract_ros_examples/src/basic_cartesian_example_node.cpp
@@ -1,0 +1,50 @@
+/**
+ * @file basic_cartesian_plan.cpp
+ * @brief Example using Trajopt for cartesian planning
+ *
+ * @author Levi Armstrong
+ * @date Dec 18, 2017
+ * @version TODO
+ * @bug No known bugs
+ *
+ * @copyright Copyright (c) 2017, Southwest Research Institute
+ *
+ * @par License
+ * Software License Agreement (Apache License)
+ * @par
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * @par
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include <tesseract_ros_examples/basic_cartesian_example.h>
+
+using namespace tesseract_ros_examples;
+
+int main(int argc, char** argv)
+{
+  ros::init(argc, argv, "basic_cartesian_example_node");
+  ros::NodeHandle pnh("~");
+  ros::NodeHandle nh;
+
+  bool plotting = true;
+  bool rviz = true;
+  int steps = 5;
+  std::string method = "json";
+
+  // Get ROS Parameters
+  pnh.param("plotting", plotting, plotting);
+  pnh.param("rviz", rviz, rviz);
+  pnh.param<std::string>("method", method, method);
+  pnh.param<int>("steps", steps, steps);
+
+  BasicCartesianExample example(nh, plotting, rviz, steps, method);
+  example.run();
+}
