+++ tesseract_ros/tesseract_examples/tesseract_ros_examples/test/puzzle_piece_auxillary_axes_example_unit.test
@@ -1,0 +1,9 @@
+<?xml version="1.0"?>
+<launch>
+  <arg name="plotting" default="false"/>
+  <arg name="rviz" default="false"/>
+  <arg name="testing" default="true"/>
+
+  <include file="$(find tesseract_ros_examples)/launch/puzzle_piece_auxillary_axes_example.launch" pass_all_args="true"/>
+  <test pkg="tesseract_ros_examples" type="tesseract_ros_examples_puzzle_piece_auxillary_axes_example_unit" test-name="puzzle_piece_auxillary_axes_example_unit"/>
+</launch>
